refactor: Split core into multiple modules
